crud

+++ controllers/campaign.controller.js
@@ -4,6 +4,7 @@ const Company = require('../models/company.model');
 const Donation = require('../models/donation.model');
 const Recruiter = require('../models/recruiter.model');
 const User = require('../models/user.model');
+const { findUserByUidWithFullPopulate, updateUser } = require('./user.controller');
 
 const findAllCampaignsWithFullPopulate = async () => {
     return await Campaign.find().populate([
@@ -65,12 +66,6 @@ const findCampaignWithFullPopulate = async (id) => {
 
 const getAllCampaigns = async (req, res) => {
     try {
-        // edit all campaigns duration
-        // let c = await Campaign.find({});
-        // await c.forEach(cc => {
-        //     cc.duration = ['01/01/2021', '01/03/2021'];
-        //     cc.save();
-        // })
         const allCampaigns = await findAllCampaignsWithFullPopulate();
         console.log("🚀 ~ file: campaign.controller.js ~ line 6 ~ getAllCampaigns ~ allCampaigns", allCampaigns);
         res.status(200).send(allCampaigns);
@@ -95,21 +90,14 @@ const getCampaignById = async (req, res) => {
 
 const createCampaign = async (req, res) => {
     try {
-        let ansCampaign;
-        // if company not exists:
-        if (!req.body.company) {
-            const newCompany = await new Company(req.body).save();
-            console.log("🚀 ~ file: campaign.controller.js ~ line 30 ~ createCampaign ~ newCompany", newCompany);
-            ansCampaign = await new Campaign({ ...req.body, company: newCompany._id }).save();
-        }
-        else {
-            ansCampaign = await new Campaign({ ...req.body }).save();
-        }
-        const user = await User.findByIdAndUpdate(req.body.userId, { $push: { campaigns: ansCampaign._id } });
-        console.log("🚀 ~ file: campaign.controller.js ~ line 66 ~ createCampaign ~ user", user)
+        const ansCampaign = await new Campaign({ ...req.body }).save();
+        const updateUser = await User.findByIdAndUpdate(req.body.userId, { $push: { campaigns: ansCampaign._id } });
+        console.log("🚀 ~ file: campaign.controller.js ~ line 95 ~ createCampaign ~ updateUser", updateUser)
         const campaign = await findCampaignWithFullPopulate(ansCampaign._id);
         console.log("🚀 ~ file: campaign.controller.js ~ line 39 ~ createCampaign ~ campaign", campaign)
-        res.status(200).send({ campaign, user });
+        const allCampaigns = await findAllCampaignsWithFullPopulate();
+        const user = await findUserByUidWithFullPopulate(updateUser.uid);
+        res.status(200).send({ campaign, user, allCampaigns });
     }
     catch (error) {
         console.log("🚀 ~ file: campaign.controller.js ~ line 31 ~ createCampaign ~ error", error);
@@ -119,13 +107,19 @@ const createCampaign = async (req, res) => {
 
 const updateCampaign = async (req, res) => {
     try {
-        const updateCampaign = await Campaign.findByIdAndUpdate(req.body._id, req.body);
-        console.log("🚀 ~ file: campaign.controller.js ~ line 50 ~ updateCampaign ~ updateCampaign", updateCampaign);
-        const campaign = await findCampaignWithFullPopulate(updateCampaign._id);
-        console.log("🚀 ~ file: campaign.controller.js ~ line 85 ~ updateCampaign ~ campaign", campaign)
-        const allCampaigns = await findAllCampaignsWithFullPopulate();
-        console.log("🚀 ~ file: campaign.controller.js ~ line 124 ~ updateCampaign ~ allCampaigns", allCampaigns)
-        res.status(200).send({ campaign, allCampaigns });
+        const checkCampaign = await Campaign.findById(req.body._id);
+        if (checkCampaign.goalRaised) {
+            throw Error('כבר תרמו לקמפיין')
+        }
+        else {
+            const updateCampaign = await Campaign.findByIdAndUpdate(req.body._id, req.body);
+            console.log("🚀 ~ file: campaign.controller.js ~ line 50 ~ updateCampaign ~ updateCampaign", updateCampaign);
+            const campaign = await findCampaignWithFullPopulate(updateCampaign._id);
+            console.log("🚀 ~ file: campaign.controller.js ~ line 85 ~ updateCampaign ~ campaign", campaign)
+            const allCampaigns = await findAllCampaignsWithFullPopulate();
+            const user = await findUserByUidWithFullPopulate(req.body.uid)
+            res.status(200).send({ campaign, allCampaigns, user });
+        }
     }
     catch (error) {
         console.log("🚀 ~ file: campaign.controller.js ~ line 45 ~ updateCampaign ~ error", error);
@@ -135,18 +129,25 @@ const updateCampaign = async (req, res) => {
 
 const deleteCampaign = async (req, res) => {
     try {
-        const deleteCampaign = await Campaign.findByIdAndDelete(req.params._id);
-        console.log("🚀 ~ file: campaign.controller.js ~ line 136 ~ deleteCampaign ~ deleteCampaign", deleteCampaign)
-        if (deleteCampaign) {
+        const checkCampaign = await Campaign.findById(req.params._id);
+        if (checkCampaign.goalRaised) {
+            throw Error('כבר תרמו לקמפיין')
+        }
+        else {
+            const deleteCampaign = await Campaign.findByIdAndDelete(req.params._id);
+            console.log("🚀 ~ file: campaign.controller.js ~ line 136 ~ deleteCampaign ~ deleteCampaign", deleteCampaign)
             // {"breed" : { $in : ["Pitbull", "Great Dane", "Pug"]}}
             const cardsToDelete = await Card.deleteMany({ _id: { $in: deleteCampaign.cards } })
             console.log("🚀 ~ file: campaign.controller.js ~ line 141 ~ deleteCampaign ~ cardsToDelete", cardsToDelete)
             const recruitersToDelete = await Recruiter.deleteMany({ _id: { $in: deleteCampaign.recruiters } })
             console.log("🚀 ~ file: campaign.controller.js ~ line 143 ~ deleteCampaign ~ recruitersToDelete", recruitersToDelete);
+            const updateUser = await User.findByIdAndUpdate(req.params.userId, { $pull: { campaigns: req.params._id } });
+            console.log("🚀 ~ file: campaign.controller.js ~ line 135 ~ deleteCampaign ~ updateUser", updateUser)
         }
         const allCampaigns = await findAllCampaignsWithFullPopulate();
-        console.log("🚀 ~ file: campaign.controller.js ~ line 150 ~ deleteCampaign ~ allCampaigns", allCampaigns)
-        res.status(200).send(allCampaigns);
+        const userUID = await User.findById(req.params.userId);
+        const user = await findUserByUidWithFullPopulate(userUID.uid)
+        res.status(200).send({ allCampaigns, user });
     } catch (error) {
         console.log("🚀 ~ file: campaign.controller.js ~ line 150 ~ deleteCampaign ~ error", error)
         res.status(500).send({ error });
